Dividi o código para facilitar o entendimento e os testes
